separate ble client class
